feat: sposta sezione abbonamento in Impostazioni
La voce "Abbonamento" è rimossa dalla navbar (desktop e mobile bottom nav,
da 5 → 4 voci) e il contenuto è integrato come sezione inline nella pagina
/dashboard/settings, seguendo il pattern industria (GitHub, Vercel, Linear).

- Rimosso link Abbonamento da layout.tsx (desktop nav)
- Rimosso tab Abbonamento da bottom-nav.tsx (+ test aggiornati a 4 voci)
- Eliminata route /dashboard/abbonamento (page + loading)
- Aggiunta card "Piano e Abbonamento" in settings/page.tsx
  (piano corrente, selezione piani, portale Stripe)
- Estratto PlanBadge come componente condiviso in components/billing/
- Aggiornati success_url/cancel_url Stripe checkout → /dashboard/settings
- Aggiornato trial-expiring.test.tsx con nuovo URL
