+++ sonar-project.properties
@@ -47,14 +47,13 @@ sonar.coverage.exclusions=\
   src/components/storico/void-receipt-dialog.tsx,\
   src/components/providers.tsx,\
   src/app/r/**/page.tsx,\
-  src/app/dashboard/abbonamento/page.tsx,\
   src/components/billing/checkout-button.tsx,\
+  src/components/billing/plan-badge.tsx,\
   src/components/settings/export-data-section.tsx,\
   src/app/dashboard/loading.tsx,\
   src/app/dashboard/cassa/loading.tsx,\
   src/app/dashboard/storico/loading.tsx,\
-  src/app/dashboard/settings/loading.tsx,\
-  src/app/dashboard/abbonamento/loading.tsx
+  src/app/dashboard/settings/loading.tsx
 
 sonar.typescript.lcov.reportPaths=coverage/lcov.info
 sonar.testExecutionReportPaths=test-report.xml
